requirements/base.txt
django==3.1.5
psycopg2==2.8.6
django-allauth==0.43.0
dj-stripe==2.4.1
django-recaptcha==2.0.6